requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Markdown==3.5.1
python-dotenv==1.0.0
Werkzeug==3.0.1
bleach==6.1.0
